Add import guard and installation instructions on line profiler for non OSX users
Test Plan:

On OSX, load up a page with the profiler turned on and change the profiling mode to line-by-line.
You should see usage instructions as normal.

Now replace the `_line_profiler.so` file with a garbage file, e.g. like this:
```
echo 'garbage' > _line_profiler.so
```
This is to emulate wrong architecture trying to be loaded.
Reload the page in browser, and you should be presented with a message similar to this once you expand the profiling results:

```
Could not load line_profiler c extension.

Try installing it like so:
  sudo pip install line_profiler==1.0b3
  (cd / && cp `python -c 'import _line_profiler; print _line_profiler.__file__'` /Users/jlfwong/khan/webapp/third_party/gae_mini_profiler)
```

Except with the paths customized for your installation.
Run the commands, reload the page, and you should see usage instructions again.