Don't disable the sampling profiler in dev_appserver
Summary:
It looks like the sampling profiler was disabled in dev mode because it didn't
work at the time, but improvements to dev_appserver since then have made this
restriction no longer necessary.

Test Plan:
Visit the logged in homepage with dev_appserver and switch to the sampling
profiler. Verify that there is data there and that it roughly matches my
understanding of what happens on login.

Reviewers: ben, chris

Reviewed By: ben

Differential Revision: http://phabricator.khanacademy.org/D6645