Improve sleeping strategy in the sampling profiler
Summary:
Previously, we were trying to achieve 250 samples per second by sleeping for 4ms
between samples, which ended up getting only about 100 samples per second, and
gave unfair weight to times where we were able to sample faster. We now figure
out the time that the next sample should theoretically take place and sleep
until then, which ensures that we get the appropriate number of samples. They're
still not distributed perfectly evenly, with occasional pauses as high as
50-100ms in extreme cases, but this should at least improve things.

Test Plan:
Run dev_appserver and load the logged in homepage with sampling enabled. Verify
that the number of samples is approximately 250 times the number of seconds
taken in the request, and that the results seem to represent the actual time
spent in the different sections of the logged in homepage.

Reviewers: ben, chris

Reviewed By: ben

Differential Revision: http://phabricator.khanacademy.org/D6647