Change the sampling profiler to display full stack traces
Summary:
The previous visualization for the sampling profiler (all frames in all
collected stack traces, sorted by frequency), wasn't very useful, except
possibly in some cases where the user knows the code very well and can spot an
unusual frame frequency.

We now send over the full stack trace for each sample and provide a way of
visualizing this information. This is useful, for example, to see what was going
on at a time when Appstats shows no RPCs active. Currently, the user picks the
sample number to view by typing it into a text box, but the plan is to change
that to a slider soon.

To avoid sending over a giant response, we compress the stack traces by keeping
a table of frame description strings and by representing a stack trace as a list
of indexes into that table.

Test Plan:
-Load the logged in home page with sampling and Appstats enabled.
-Try various valid sample numbers and make sure the changes are snappy and the
stack traces themselves look reasonable.
-Cross reference with Appstats by typing in the sample number corresponding to
an RPCless time (assuming 250 samples per second) and making sure that it's
doing what I thought it was doing.
-Try out-of-range integers, non-integers and make sure they clear the table
without any errors showing up in the JS console.

Reviewers: ben, chris

Reviewed By: ben

Differential Revision: http://phabricator.khanacademy.org/D6648