Change the sampling profile view to use a slider
Summary:
Instead of having to type in a sample number, you can now drag the slider to the
right place in time.

Also change the stack trace table to not change in size when dragging the slider around.

Test Plan:
Load the logged-in homepage and try out the slider. Make sure the results are
pretty much the same as with the text input and that the rendering is
responsive. Load the logged-out homepage and make sure that the slider shows up
there too and that it gives reasonable results.

Reviewers: chris

Reviewed By: chris

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6651