Don't lose sort order in the sampling profile viewer
Summary:
Previously, the tableSorter wouldn't re-sort itself when the contents were
updated, and I ended up needing to do some hacks to make it work. The intention
here was to make it easy to watch the stack trace change over time as a way of
making it visually obvious how much time is spent in each function, but
unfortunately, that doesn't quite work yet, since there's enough framework code
at the top of the stack that the important frames are out of view.

I also split up updateSampleNumber into two functions, since it got too long.

Test Plan:
Go to the logged-in homepage in dev_appserver and make sure that the sampling
profiler results look reasonable. Sort the depth in reverse order and make sure
that the results remain sorted in reverse order as I change the sample number,
and make sure that the table doesn't flicker too much.

Reviewers: ben, chris

Reviewed By: ben

Differential Revision: http://phabricator.khanacademy.org/D6650