Allow filtering of frames and add help message to the sampling profiler
Summary:
Add a new "Hide frames" slider that hides some number of top-level frames,
which is useful when viewing the frames in ascending order of depth to watch
the stack change over time (since without this feature, the view would be
filled with stack frames that are the same for the entire request). I
considered trying to automatically filter out these frames, but the obvious
approach would only filter out half the frames that really should be filtered
out, so I think it makes more sense to make it user-specifiable.

Also add a "how to use" message that gives some intended use cases.

Test Plan:
On both the logged in and logged out homepage:
-Load the page and make sure the profile works like it did before.
-Drag the "hide frames" slider to various values and make sure that no errors
appear and that the stack trace changes in a smooth way when viewed in
ascending order of depth.

Reviewers: ben, chris

Reviewed By: chris

CC: marcos

Differential Revision: http://phabricator.khanacademy.org/D6654