Add option to sampling profiler to additionally sample memory.
Summary:
This adds an option to additionally sample memory.  Since memory sampling is
slow on prod, it does it fairly infrequently.  (After some further profiling,
we may be able to turn it up a bit.)

Currently it does nothing on a dev server, since the dev server does not allow
importing the python module, and does not return useful data from the prod
runtime API.

Screenshots:
{F23421}
{F23422}

Test Plan:
Set the dev server fallback to return random data instead of zero, to check
that the interface work.  Test the profiler in the dev shell.  Will also test
on staging.

Reviewers: alan

Reviewed By: alan

Differential Revision: http://phabricator.khanacademy.org/D10320