Fixes two little bugs in sampling memory profiler.
The display was breaking if memory sampling was off because I was using jQuery
incorrectly, and I forgot to add the new option in one place.

Test Plan:
Load a page with the miniprofiler, with memory sampling both on and off.  Check
that the profile renders correctly in both cases.

Auditors: alan