Make mini profiler detail window float at the top of the page
Summary:
Previously if you had a really long list of AJAX requests in the mini profiler,
you would have to scroll down, click on one, and then scroll way up to the
detail window.  This way, the detail window will float down the page with you.
{F33796}

Test Plan: Load a page, get it to make lots of AJAX requests.

Reviewers: alan

Reviewed By: alan

Differential Revision: http://phabricator.khanacademy.org/D11272