Add sampling profiler method to dump to .cpuprofile format.
Summary:
It's currently not hooked up to anything, but if you run the sampling profiler
manually, write Profiler.cpuprofiler_results() to a file, and upload that to
Chrome, it should work.
{F40100}

Known issue: the very last sample gets displayed a bit weirdly.  I'll fix this if I
have time, but it's not a huge issue since in real things we want to optimize
we'll probably have tens or hundreds of samples.

Test Plan:
Profiled time.sleep() from the dev shell, loaded the profile into chrome,
checked that things looked reasonable.

Reviewers: alan

Reviewed By: alan

Differential Revision: http://phabricator.khanacademy.org/D11844