Drop instrumented profile from the default mode
Summary:
Enabling the instrumented profile slows the request significantly and in
usage I've observed it's not worth having on by default in either
development or production. The RPC trace is important enough often
enough to have on by default, it's often the case that a repeat of the
same request has a different RPC trace due to server-side caching.

Test Plan:
Loaded a page with mini profiler and no cookie, saw that I got an RPC
trace but not an instrumented profile. Loaded the "Settings" pane and
say that "CPU profiling" is disabled and "RPC profiling" is enabled.

Reviewers: alpert, desmond

Reviewed By: desmond

Differential Revision: http://phabricator.khanacademy.org/D13413