Integrate an embedded sampling profile viewer from Chrome DevTools
Summary:
With this, the flame chart visualization of a sampled profile is one
click away and there's no need to open Chrome DevTools to see it.

It only works in Chrome, so for other browsers we display unlinked text
and a tooltip within the mini profiler.

The license for the included DevTools code from the chromium project
lives in static/chrome/inspector-20140603/devtools.html

Test Plan:
Loaded up mini profiler in Chrome on the KA homepage, clicked the
"Visualize" link in the mini profiler's profile view, and saw the new
tab open with a flame chart visualization.

Then I loaded up mini profiler in Safari and saw the unlinked text
"Visualize in Chrome".

Reviewers: desmond, jlfwong

Reviewed By: jlfwong

Subscribers: alan, ben, marcos

Differential Revision: http://phabricator.khanacademy.org/D13549