Add a collapse button to gae_mini_profiler
Summary:
Some people don't like to see the mini profiler in their face all the time. It can also impede UI iteration since it overlays things. So, I added this button that allows you to collapse the widget. The "hidden" setting persists through page refreshes via a cookie.

I did my best on "designing" this button to be tiny, but I'm not really happy with it, so I'm very open to CSS/HTML tips on making it more usable.

Test Plan:
Load up any page locally, see the mini profiler with the new collapse button, click it
Notice the entries are hidden.
Trigger some more RPCs on the page (hover over the learn menu), notice the entries are still hidden
Click the open button, to display the entries again. Everything behaves as normal.
Click to close, hard-refresh the page. Notice that the entries are still collapsed

Reviewers: csilvers, alan

Subscribers: jace, alpert

Differential Revision: http://phabricator.khanacademy.org/D14150