Support memcache chunking in the sampling profiler.
Summary:
Now we'll be able to show sampled profile results even if the
pickled+compressed output is >1M.

Test Plan:
Deployed a znd that used this new code, and used the sampling profiler
on the homepage of a new request to see how long it took
get_path_table() to run (I deleted the `@layer_cache` for
get_path_table() first).  I saw the mini-profiler dropdown appear,
whereas before it didn't, and could visualize the profile.

Reviewers: chris, benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D30226