Add an icon for error logs to the mini profiler
Summary:
The mini profiler has a bunch of little icons when various potentially-bad
things happen.  But an error log have one -- you have to open up the request to
have any hint of it!  Now it will -- we'll show a little orange thingy for
errors and a red one for criticals.

Screenshot: {F594796}

Test Plan:
Substate into webapp, and edit `RequestHandler.dispatch` to log an error always
and a critical sometimes randomly, and see things like the above screenshot.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D32513