Allow sampling profile cpuprofiles to be imported into speedscope.app
Summary:
It turns out that speedscope.app already supports our cpuprofile format
(which appears to be an old format used by the v8 profiler), we just had
a bug in our profile output that caused it to crash on import. By
removing the addition in this patch, the profiles import into speedscope
without an issue!

I couldn't find any git history to explain why we needed this addition
in the first place, but I don't think it's necessary. The existing
chrome visualizer still works and the documentation of the new v8
profile format [1] seems to indicate that the timestamps should be
relative to the `startTime` so no addition should be necessary.

I also add a few comments in this patch to help document this old
format.

[1]
https://github.com/v8/v8/blob/44bd8fd7/src/inspector/js_protocol.json#L1485

Test Plan:
Substate this patch into webapp, load a homepage and open the mini
profiler.

Verified that the 'Visualize' button continues to display the flamechart
as it did before.

Downloaded the cpuprofiler, then imported it into the interface at
speedscope.app. Before this patch, a JS error would occur. After this
patch, no error and pretty flamecharts!

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D53400