Replace old chrome flamechart viewer with speedscope
Summary:
This patch replaces the embedded chrome flamechart viewer with
speedscope [1].

Speedscope offers a number of benefits over the old viewer:
 - When you click on a stack frame, you can now see things the filename
   where the function resides, rather than grepping for a matching string.
 - We can now easily export stackprofiles and have other engineers
   import them to examine later (this was somewhat possible with the
   current chrome viewer, but it wasn't well supported.)
 - The UI offers new 'Left-Heavy' and 'Sandwich' visualization modes
   (details here [2])
 - It's much more intuitive to zoom into particular areas of the trace

The code within static/speedscope is vendored directly without any
modifications. I downloaded the v1.5.0 release from github [3] and moved
the unzipped directory there.

Speedscope supports loading a profile by url, though it requires an
absolute url so I needed to do a bit of url munging to ensure that it
will work locally as well as in production.

Since we no longer use the chrome viewer, I remove it entirely in this
patch and upate the reference in the README to point to Speedscope
instead.

[1] https://www.speedscope.app/
[2] http://jamie-wong.com/post/speedscope/
[3] https://github.com/jlfwong/speedscope/releases

Test Plan:
Substate this patch into webapp, load a homepage and open the mini
profiler.

Clicked 'Visualize' within the sampling profiler:

{F1413683}

Before this patch, we would see the chrome profiler:

{F1413682}

After this patch, we see the new shiny speedscope ui:

{F1413681}

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D53401