Fix invisible log slider
Summary:
Some time ago, the slider that allows you to configure the log level
disappeared. It would still work if you a) knew that the bar was
clickable and b) clicked on just the right area. But that's not the most
intuitive interface, so this patch makes it visible again.

I couldn't figure out the correct css styles to make this old version of
jquery slider work, so I replaced it with a plain html range input
instead, which seems to work just as well.

Test Plan:
Substate this patch in webapp, load the homepage and open the logs area
of the mini profiler.

Before this patch, you couldn't see the logs level slider:

{F1414238}

After this patch the slider appears and continues to work

{F1414237}

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D53411