Capture and display stackdriver trace details in RPC profile area
Summary:
Stackdriver trace can be an useful tool for folks using the mini
profiler. This patch captures the trace id if one is available and
provides a link to the relevant part of the stackdriver UI.

I also add a help popup to explain why each request may not have a trace
associated with it.

Test Plan:
Locally I used a chrome extension [1] to automatically add the
'X-Cloud-Trace-Context' header with a dummy value to every request I made.

I substated this patch into webapp and loaded the homepage.

I verified that the new link appears if the header is set:

{F1414541}

and that text appears if no header is set:

{F1414543}

and finally that the help popup shows up:

{F1414542}

I also deployed to a znd and verified that the links actually work with
valid data.

[1]
https://chrome.google.com/webstore/detail/modify-headers-for-google/innpjfdalfhpcoinfnehdnbkglpmogdi/related

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D53426