Provide an explicit default for the cloud trace header
Normally, `os.environ.get()` defaults to `''` instead of `None` like it
would for a normal dict.  But in tests we sometimes mock it to a copy of
itself, which is a normal dict and defaults to `None`.  Now, we provide
an explicit default so this doesn't crash.

Test plan:
in webapp, `tools/runtests.py dev/extbackup/tests/main_test.py`

Auditors: dhruv